+++ projectroles/templatetags/projectroles_tags.py
@@ -62,6 +62,12 @@ def get_project_list_indent(project, list_parent):
 
 
 @register.simple_tag
+def get_project_search_list(user, title):
+    """Return flat project list based on a title string search"""
+    return Project.objects.find_by_full_title(user, title, type='PROJECT')
+
+
+@register.simple_tag
 def omics_constant(value):
     """Get value from OMICS_CONSTANTS in settings"""
     return OMICS_CONSTANTS[value] \